Hack in python script to change chromosome output by msprime from 1 to the one the user specified in VCF output.
Also something seemed incorrect, at least when run as a docker container as opposed
to on precisionFDA because the output was not being placed in the intended shared/
directory where it would be visible to the host, and getting lost with the
container on exit. This would not happen in pFDA because there is no container
and the file is uploaded explicitly in the pFDA script.

Need to test that this will now still work on pFDA